requirements.txt
gpt-index
html2text
python-dotenv